requirements.txt
bs4
httpx
fastapi
pymemcache
pycryptodome
cryptography
aiohttp
btoa
